sszutils: restore ErrPlatformOverflowFn's released signature
The value parameter was narrowed to uint64, which does not compile for a
caller passing an int, a uint32 or an int64. sszutils is inside the
compatibility scope and this is a minor release, so the parameter is any
again, as v1.3.2 declared it.

The width a value states still decides which limit it passed: an integer
wider than the SSZ offset is past every target's reach, not only this one's.
A value of no integer type states no width and is reported against the
platform's range.

+++ dynssz.go
@@ -431,7 +431,7 @@ func (d *DynSsz) MarshalSSZTo(source any, buf []byte, opts ...CallOption) ([]byt
 	// is no wider is the whole int range: the bytes already in buf then take
 	// the sum past it.
 	if size > int64(math.MaxInt)-int64(len(buf)) {
-		return nil, sszutils.ErrPlatformOverflowFn("SSZ size", uint64(size))
+		return nil, sszutils.ErrPlatformOverflowFn("SSZ size", size)
 	}
 	needed := len(buf) + int(size)
 	if cap(buf) < needed {
